♻️ Refactor iscn signing function
