+++ util/CosmosHelper.js
@@ -1,18 +1,15 @@
 import BigNumber from 'bignumber.js';
 import {
   COSMOS_CHAIN_ID,
-  ISCN_TESTNET_CHAIN_ID,
   COSMOS_DENOM,
 } from '@/constant';
 import { timeout } from '@/util/misc';
-import { MsgCreateISCN } from './cosmos/iscn';
 
 export const DEFAULT_GAS_PRICE = [{ amount: 1000, denom: 'nanolike' }];
 export const DEFAULT_ISCN_GAS_PRICE = [{ amount: 0, denom: 'nanolike' }];
 
 let Cosmos;
 let api;
-let iscnApi;
 
 async function initCosmos() {
   if (api) return;
@@ -21,15 +18,6 @@ async function initCosmos() {
   ]));
   if (Cosmos.default) Cosmos = Cosmos.default;
   api = new Cosmos('/api/cosmos/lcd', COSMOS_CHAIN_ID);
-}
-
-async function initISCNCosmos() {
-  if (iscnApi) return;
-  ([Cosmos] = await Promise.all([
-    import(/* webpackChunkName: "web3" */ '@lunie/cosmos-api'),
-  ]));
-  if (Cosmos.default) Cosmos = Cosmos.default;
-  iscnApi = new Cosmos('/api/cosmos/iscn-dev/lcd', ISCN_TESTNET_CHAIN_ID);
 }
 
 function LIKEToNanolike(value) {
@@ -225,36 +213,4 @@ export function transferMultiple({
   const amounts = values.map(v => LIKEToAmount(v));
   const msgPromise = MsgSendMultiple(from, { toAddresses: tos, amounts });
   return sendTx(msgPromise, signer, { memo, simulate });
-}
-
-export async function signISCNPayload({
-  userId,
-  displayName,
-  cosmosWallet,
-  fingerprint,
-  title,
-  tags = [],
-  type = 'article',
-  license,
-  publisher,
-  memo,
-}, signer, { simulate = false } = {}) {
-  if (!iscnApi) await initISCNCosmos();
-  const msgPromise = MsgCreateISCN(iscnApi,
-    {
-      id: userId,
-      displayName,
-      cosmosWallet,
-    },
-    {
-      fingerprint,
-      title,
-      tags,
-      type,
-    },
-    {
-      license,
-      publisher,
-    });
-  return sendTx(msgPromise, signer, { memo, simulate, gasPrices: DEFAULT_ISCN_GAS_PRICE });
 }
